Refactor to remove separate farm_public endpoints. Related to #68
